v3.6.0 — Dalamud API 15
Bump Dalamud.NET.Sdk + DalamudPackager to 15.0.0. Single source change
required: IClientState.TerritoryChanged delegate widened from
Action<ushort> to Action<uint>. Cast at the boundary; pluginState.territory
stays ushort.

+++ VenueManager/Plugin.cs
@@ -570,10 +570,10 @@ namespace VenueManager
     public void ToggleConfigUI() => MainWindow.Toggle();
     public void ToggleMainUI() => MainWindow.Toggle();
 
-    private unsafe void OnTerritoryChanged(ushort territory)
+    private unsafe void OnTerritoryChanged(uint territory)
     {
       // Save current user territory
-      pluginState.territory = territory;
+      pluginState.territory = (ushort)territory;
 
       bool inHouse = false;
       try
